requirements.txt
google-cloud-storage
google-cloud-bigquery
requests
pandas
python-dotenv
jupyterlab
